1. Move the delete button to right-top corner. 2. Show a confirmation dialog before delete. 3. Only admin can delete clauses

+++ DeviceAdministration/Web/Scripts/Views/Device/DeviceFilter.js
@@ -404,12 +404,18 @@
                     clause.checked(!clause.checked());
                     self.model.isChanged(true);
                 },
-                remove: function () {
-                    api.deleteSuggestedClause(self.model.getClauseModel(clause), function () {
-                        getSuggestedClauses();
+                remove: function (data, e) {
+                    IoTApp.Helpers.Dialog.confirm(resources.deleteClauseConfirmation, function (result) {
+                        if (result) {
+                            api.deleteSuggestedClause(self.model.getClauseModel(clause), function () {
+                                getSuggestedClauses();
+                            });
+
+                            self.model.removeClause(clause);
+                        }
                     });
 
-                    self.model.removeClause(clause);
+                    e.stopPropagation();
                 }
             };
